+++ src/App.tsx
@@ -7,11 +7,12 @@ import Counter from './pages/counter/counter';
 import { Calcul } from './calcul/calcul';
 import TaskPage from './pages/TaskPage/TaskPage';
 import TextPage from './pages/storybook/componentsPage/TextPage/TextPage';
+import TaskTablePage from './pages/TaskTablePage/TaskTablePage';
 
-type PageTypes = 'storybook' | 'counter' | 'calculate_(Button)' | 'myCalculate_(input)' | 'taskPage'
+type PageTypes = 'storybook' | 'counter' | 'calculate_(Button)' | 'myCalculate_(input)' | 'taskPage' | 'TaskTable'
 
 const App: FC = () => {
-  const [page, setPage] = useState<PageTypes>('storybook');
+  const [page, setPage] = useState<PageTypes>('TaskTable');
   
   const [isDarkTheme, setIsDarkTheme] = useState<boolean>(false);
 
@@ -28,11 +29,13 @@ const App: FC = () => {
       case 'calculate_(Button)': return <Calcul/>;
       case 'myCalculate_(input)': return null;
       case 'taskPage': return <TaskPage/>;
+      case 'TaskTable': return <TaskTablePage headerlist={[]}/>;
     }
   }
 
   const renderNavigation = () => (
     <div className='navigation'>
+      <Button label='Table' onClick={() => setPage('TaskTable')}/>
       <Button label='storybook' onClick={() => setPage('storybook')}/>
       <Button label='counter' onClick={() => setPage('counter')}/>
       {/* <Button label='myCalculate (input)' onClick={() => setPage('myCalculate (input)')}/> */}
