+++ src/components/TaskTable/TaskTableComponents/TaskTableRow/TaskTableRow.css
@@ -1,0 +1,15 @@
+/**
+ * основной стиль
+ */
+.Table {
+  display: flex;
+  flex-direction: column;
+
+  width: 100%;
+}
+/**
+ * Стиль строки
+ */
+.tableRow:last-child {
+  border-bottom: 1px solid black;
+}
