+++ src/pages/TaskTablePage/index.ts
@@ -1,0 +1,3 @@
+import TaskTablePage from "./TaskTablePage";
+
+export default TaskTablePage;
